+++ src/vmware/azext_vmware/aaz/latest/vmware/script_execution/_delete.py
@@ -19,13 +19,13 @@ class Delete(AAZCommand):
     """Delete a ScriptExecution in a private cloud
 
     :example: Delete a script execution.
-        az vmware script-execution delete --resource-group group1 --private-cloud cloud1 --name addSsoServer
+        az vmware script-execution delete --resource-group group1 --private-cloud cloud1 --script-execution-name addSsoServer
     """
 
     _aaz_info = {
-        "version": "2023-09-01",
+        "version": "2024-09-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.avs/privateclouds/{}/scriptexecutions/{}", "2023-09-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.avs/privateclouds/{}/scriptexecutions/{}", "2024-09-01"],
         ]
     }
 
@@ -159,7 +159,7 @@ class Delete(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-09-01",
+                    "api-version", "2024-09-01",
                     required=True,
                 ),
             }